pyproject.toml
[project]
name = "social-media-sdk"
version = "0.1.0"
description = "A production-ready Python SDK for social media data extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*", "models*", "platforms*"]
